refactor: move service into lib and add build_service() for in-process hosting
Move GeocodeService, its McpService impl, server_config(), and the error/arg
helper fns out of the binary and into a new src/service.rs module. Re-export
GeocodeService and server_config from the library and add a build_service()
entry point so the service can be linked directly into a client binary instead
of only being reachable by launching the CLI. main.rs is now a thin wrapper that
constructs via the library and keeps the existing mcp_core::run_simple launch.
